moved restore cmd in the dgraph/cmd dir.
